requirements.txt
Flask==3.0.0
stravalib==2.4
python-dotenv==1.0.0
bcrypt==4.1.2
Flask-Login==0.6.3
mcp>=1.0.0

# Production
gunicorn==21.2.0
